ai_study_assistant/requirements.txt
streamlit
python-dotenv
pypdf
python-pptx
faiss-cpu
sentence-transformers
openai
google-cloud-translate
langchain-community
langchain-core
langchain-text-splitters